Function hotg_rune_cli::run::multi::new_multiplexer

pub fn new_multiplexer<S, I>(
    sources: I
) -> impl Fn() -> Result<Box<dyn Capability>, Error> where
    I: IntoIterator<Item = S::Source>,
    S: SourceBackedCapability, 

Get a function for creating new Capability objects which can be initialized by switching between different source values based on the "source" parameter (i.e. an image or audio clip).